+++ src/lib/index.ts
@@ -60,4 +60,5 @@ export {
 
 export { splitPath } from './stringPath.js';
 
-export type { ErrorStatus } from './utils.js';
+export type { ErrorStatus, MergeUnion, MergeFormUnion } from './utils.js';
+export { mergeFormUnion } from './utils.js';